LPS-80848 portal-search-web: New Search Portlets: rename packages following portlet.shared.search convention
https://issues.liferay.com/browse/LPS-80848
