+++ modules/apps/portal-search/portal-search-web/src/main/java/com/liferay/portal/search/web/internal/site/facet/portlet/shared/search/SiteFacetPortletSharedSearchContributor.java
@@ -12,11 +12,14 @@
  * details.
  */
 
-package com.liferay.portal.search.web.internal.site.facet.portlet;
+package com.liferay.portal.search.web.internal.site.facet.portlet.shared.search;
 
 import com.liferay.portal.kernel.search.facet.Facet;
 import com.liferay.portal.kernel.search.facet.ScopeFacetFactory;
 import com.liferay.portal.search.web.internal.site.facet.constants.SiteFacetPortletKeys;
+import com.liferay.portal.search.web.internal.site.facet.portlet.ScopeFacetBuilder;
+import com.liferay.portal.search.web.internal.site.facet.portlet.SiteFacetPortletPreferences;
+import com.liferay.portal.search.web.internal.site.facet.portlet.SiteFacetPortletPreferencesImpl;
 import com.liferay.portal.search.web.portlet.shared.search.PortletSharedSearchContributor;
 import com.liferay.portal.search.web.portlet.shared.search.PortletSharedSearchSettings;
